+++ app/src/main/java/org/sopt/dosopttemplate/data/model/remote/request/auth/RequestLoginDto.kt
@@ -1,14 +1,12 @@
-package org.sopt.dosopttemplate.data.model.remote.model.request.auth
+package org.sopt.dosopttemplate.data.model.remote.request.auth
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestSignupDto(
+data class RequestLoginDto(
     @SerialName("username")
     val username: String,
     @SerialName("password")
     val password: String,
-    @SerialName("nickname")
-    val nickname: String,
 )
