+++ app/src/main/java/org/sopt/dosopttemplate/data/service/HomeService.kt
@@ -1,10 +1,10 @@
 package org.sopt.dosopttemplate.data.service
 
-import org.sopt.dosopttemplate.data.model.remote.model.response.follower.ResponseFollowerDto
+import org.sopt.dosopttemplate.data.model.remote.response.follower.ResponseFollowerDto
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-interface FollowerService {
+interface HomeService {
     @GET("api/users")
     suspend fun getFollowerList(
         @Query("page") page: Int
