+++ app/src/main/java/org/sopt/dosopttemplate/presentation/auth/LoginViewModel.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import org.sopt.dosopttemplate.data.model.local.UserInfo
 import org.sopt.dosopttemplate.data.model.remote.api.ServicePool
-import org.sopt.dosopttemplate.data.model.remote.model.request.auth.RequestLoginDto
+import org.sopt.dosopttemplate.data.model.remote.request.auth.RequestLoginDto
 import org.sopt.dosopttemplate.util.UiState
 
 class LoginViewModel : ViewModel() {
